+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
 
 ### Changed
 
+- 🧹 (frontend) Refactor SettingsDialogExtended layout styles (Closes #30)
 - 🧹 (infra) Centralized docker compose `--user` handling (Closes #29)
 - 🔒️ (infra) Avoid writing Vaultwarden credentials (Closes #28)
 - ♿️(frontend) Refactor formatDate replace-chain #27